fix a copy-paste error

+++ test/stb_image_test.exs
@@ -29,9 +29,9 @@ defmodule StbImageTest do
 
   test "decode as f32" do
     {:ok, img, shape, type, channels} =
-      StbImage.from_file(Path.join(__DIR__, "test.png"), 0, :u16)
+      StbImage.from_file(Path.join(__DIR__, "test.png"), 0, :f32)
 
-    assert type == :u16
+    assert type == :f32
     assert shape == {2, 3, 4}
     assert channels == :rgba
